// packages/photon-grid-core/src/types/grid.types.d.ts
export interface CellRange {
    startRowIndex: number;
    endRowIndex: number;
    startColIndex: number;
    endColIndex: number;
}